backend/requirements.txt
# Digital Force Backend — Python 3.14 + Windows compatible
# Using flexible (>=) pins for packages that need 3.14 wheels.
# --only-binary=:all: ensures zero compilation.

# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0

# Config & Validation (pydantic 2.11+ has Python 3.14 wheels)
pydantic>=2.11.0
pydantic-settings>=2.6.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.1
cryptography>=43.0.3

# AI / Agent Layer (all pure Python wheels)
langchain>=0.3.13
langchain-core>=0.3.28
langchain-groq>=0.2.3
langgraph>=0.2.60
groq>=0.13.0
httpx>=0.27.0

# RAG / Vector DB (pure Python)
qdrant-client>=1.12.1

# Document Parsing (pre-built binary wheels)
pymupdf>=1.25.1
python-docx>=1.1.2

# HTML Parsing
beautifulsoup4>=4.12.3

# HTTP / Publishing
requests>=2.32.3

# Email templates
jinja2>=3.1.4

# File Operations
aiofiles>=24.1.0

# Utilities (all pure Python)
python-dotenv>=1.0.1
python-slugify>=8.0.4
arrow>=1.3.0
tenacity>=9.0.0
structlog>=24.4.0
